+++ src/main/java/bridge/bridge_bank/domain/accounting_validation/AccountingValidationService.java
@@ -2,7 +2,7 @@ package bridge.bridge_bank.domain.accounting_validation;
 
 import bridge.bridge_bank.domain.ledger.entity.LedgerEntryType;
 import bridge.bridge_bank.domain.ledger.repository.LedgerQueryRepository;
-import bridge.bridge_bank.domain.transfer_transaction.TransferTransactionResultQueryRepository;
+import bridge.bridge_bank.domain.transfer_transaction.repository.TransferTransactionResultQueryRepository;
 import bridge.bridge_bank.domain.transfer_transaction.entity.TransferTransactionType;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
