+++ src/main/java/bridge/bridge_bank/domain/reserve_transfer_schedule/repeat/ReserveRepeatTransferScheduleTargetOption.java
@@ -1,4 +1,4 @@
-package bridge.bridge_bank.domain.transfer.reserve.repeat;
+package bridge.bridge_bank.domain.reserve_transfer_schedule.repeat;
 
 import lombok.Getter;
 import lombok.Setter;
