+++ src/main/java/bridge/bridge_bank/domain/reserve_transfer_schedule/repeat/repository/ReserveRepeatTransferScheduleQueryRepository.java
@@ -1,8 +1,8 @@
-package bridge.bridge_bank.domain.transfer.reserve.repeat.repository;
+package bridge.bridge_bank.domain.reserve_transfer_schedule.repeat.repository;
 
-import bridge.bridge_bank.domain.transfer.reserve.repeat.QReserveRepeatTransferSchedule;
-import bridge.bridge_bank.domain.transfer.reserve.repeat.entity.ReserveRepeatTransferSchedule;
-import bridge.bridge_bank.domain.transfer.reserve.repeat.ReserveRepeatTransferScheduleTargetOption;
+import bridge.bridge_bank.domain.transfer.reserve_transfer_schedule.repeat.QReserveRepeatTransferSchedule;
+import bridge.bridge_bank.domain.reserve_transfer_schedule.repeat.entity.ReserveRepeatTransferSchedule;
+import bridge.bridge_bank.domain.reserve_transfer_schedule.repeat.ReserveRepeatTransferScheduleTargetOption;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
