+++ src/main/java/bridge/bridge_bank/domain/reserve_transfer_schedule/repeat/repository/ReserveRepeatTransferScheduleRepository.java
@@ -1,6 +1,6 @@
-package bridge.bridge_bank.domain.transfer.reserve.repeat.repository;
+package bridge.bridge_bank.domain.reserve_transfer_schedule.repeat.repository;
 
-import bridge.bridge_bank.domain.transfer.reserve.repeat.entity.ReserveRepeatTransferSchedule;
+import bridge.bridge_bank.domain.reserve_transfer_schedule.repeat.entity.ReserveRepeatTransferSchedule;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReserveRepeatTransferScheduleRepository extends JpaRepository<ReserveRepeatTransferSchedule, Long> {
