+++ src/main/java/bridge/bridge_bank/domain/transfer_transaction/repository/TransferTransactionResultQueryRepository.java
@@ -1,4 +1,4 @@
-package bridge.bridge_bank.domain.transfer_transaction;
+package bridge.bridge_bank.domain.transfer_transaction.repository;
 
 import bridge.bridge_bank.domain.transfer_transaction.entity.QTransferTransactionResult;
 import bridge.bridge_bank.domain.transfer_transaction.entity.TransferTransactionType;
