+++ src/main/java/bridge/bridge_bank/domain/transfer_transaction/repository/TransferTransactionResultRepository.java
@@ -1,4 +1,4 @@
-package bridge.bridge_bank.domain.transfer_transaction;
+package bridge.bridge_bank.domain.transfer_transaction.repository;
 
 import bridge.bridge_bank.domain.transfer_transaction.entity.TransferTransactionResult;
 import org.springframework.data.jpa.repository.JpaRepository;
