+++ src/test/java/bridge/bridge_bank/domain/transfer/reserve_transfer_schedule/repeat/ReserveRepeatTransferScheduleServiceTest.java
@@ -1,9 +1,11 @@
-package bridge.bridge_bank.domain.transfer.reserve.repeat;
+package bridge.bridge_bank.domain.transfer.reserve_transfer_schedule.repeat;
 
-import bridge.bridge_bank.domain.transfer.reserve.repeat.entity.RepeatType;
-import bridge.bridge_bank.domain.transfer.reserve.repeat.entity.ReserveRepeatTransferSchedule;
-import bridge.bridge_bank.domain.transfer.reserve.repeat.repository.ReserveRepeatTransferScheduleQueryRepository;
-import bridge.bridge_bank.domain.transfer.reserve.repeat.repository.ReserveRepeatTransferScheduleRepository;
+import bridge.bridge_bank.domain.reserve_transfer_schedule.repeat.ReserveRepeatTransferScheduleService;
+import bridge.bridge_bank.domain.reserve_transfer_schedule.repeat.ReserveRepeatTransferScheduleTargetOption;
+import bridge.bridge_bank.domain.reserve_transfer_schedule.repeat.entity.RepeatType;
+import bridge.bridge_bank.domain.reserve_transfer_schedule.repeat.entity.ReserveRepeatTransferSchedule;
+import bridge.bridge_bank.domain.reserve_transfer_schedule.repeat.repository.ReserveRepeatTransferScheduleQueryRepository;
+import bridge.bridge_bank.domain.reserve_transfer_schedule.repeat.repository.ReserveRepeatTransferScheduleRepository;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
